Proposed changes to file structure
